refactor: reuse PgWallet for MWST as profiles strategy
