+++ acapy_wallet_upgrade/pg_connection.py
@@ -176,18 +176,33 @@ class PgConnection(DbConnection):
             self._conn = None
 
     def get_wallet(self) -> "PgWallet":
-        return PgWallet(self._conn)
+        return PgWallet(self._conn, self._conn, "items_old", None)
 
 
 class PgWallet(Wallet):
-    def __init__(self, conn: asyncpg.Connection):
-        self._conn = conn
+    def __init__(
+        self,
+        old_conn: asyncpg.Connection,
+        new_conn: asyncpg.Connection,
+        items_table: str,
+        wallet_id: str,
+    ):
+        self._old_conn = old_conn
+        self._new_conn = new_conn
+        self._items_table = items_table
+        self._wallet_id = wallet_id
+        self._profile_id = None
+
+    @property
+    def profile_id(self):
+        if not self._profile_id:
+            raise UpgradeError("Profile has not been initialized")
+        return self._profile_id
 
     async def insert_profile(self, name: str, key: bytes):
         """Insert the initial profile."""
-        async with self._conn.transaction():
-
-            await self._conn.execute(
+        async with self._new_conn.transaction():
+            id_row = await self._new_conn.fetch(
                 """
                     INSERT INTO profiles (name, profile_key) VALUES($1, $2)
                     ON CONFLICT DO NOTHING RETURNING id
@@ -195,9 +210,16 @@ class PgWallet(Wallet):
                 name,
                 key,
             )
+            self._profile_id = id_row[0][0]
+            return self._profile_id
 
     async def get_metadata(self):
-        stmt = await self._conn.fetch("SELECT value FROM metadata")
+        if self._wallet_id:
+            stmt = await self._old_conn.fetch(
+                "SELECT value FROM metadata WHERE wallet_id = $1", (self._wallet_id)
+            )
+        else:
+            stmt = await self._old_conn.fetch("SELECT value FROM metadata")
         found = None
         if stmt != "":
             for row in stmt:
@@ -212,47 +234,56 @@ class PgWallet(Wallet):
             raise Exception("Row not found")
 
     async def fetch_pending_items(self, limit: int):
-        """Fetch un-updated items."""
+        """Fetch un-updated items by wallet_id, if it exists."""
         while True:
-            stmt = await self._conn.fetch(
-                """
-                SELECT i.id, i.type, i.name, i.value, i.key,
-                (SELECT string_agg(encode(te.name::bytea, 'hex') || ':' || encode(te.value::bytea, 'hex')::text, ',')
-                    FROM tags_encrypted te WHERE te.item_id = i.id) AS tags_enc,
-                (SELECT string_agg(encode(tp.name::bytea, 'hex') || ':' || encode(tp.value::bytea, 'hex')::text, ',')
-                    FROM tags_plaintext tp WHERE tp.item_id = i.id) AS tags_plain
-                FROM items_old i LIMIT $1;
-                """,  # noqa
-                limit,
-            )
-            if not stmt:
+            command = """
+                    SELECT i.id, i.type, i.name, i.value, i.key,
+                    (SELECT string_agg(encode(te.name::bytea, 'hex') || ':' || encode(te.value::bytea, 'hex')::text, ',')
+                        FROM tags_encrypted te WHERE te.item_id = i.id) AS tags_enc,
+                    (SELECT string_agg(encode(tp.name::bytea, 'hex') || ':' || encode(tp.value::bytea, 'hex')::text, ',')
+                        FROM tags_plaintext tp WHERE tp.item_id = i.id) AS tags_plain
+                    """  # noqa
+            if self._wallet_id:
+                command += (
+                    f"FROM {self._items_table} i WHERE i.wallet_id = $2 LIMIT $1;"
+                )
+                rows = await self._old_conn.fetch(
+                    command,
+                    limit,
+                    self._wallet_id,
+                )
+            else:
+                command += f"FROM {self._items_table} i LIMIT $1;"
+                rows = await self._old_conn.fetch(command, limit)
+            if not rows:
                 break
-            yield stmt
+            yield rows
 
     async def update_items(self, items):
         """Update items in the database."""
         del_ids = []
         for item in items:
             del_ids = item["id"]
-            async with self._conn.transaction():
-                ins = await self._conn.fetch(
+            async with self._new_conn.transaction():
+                ins = await self._new_conn.fetch(
                     """
                         INSERT INTO items (profile_id, kind, category, name, value)
-                        VALUES (1, 2, $1, $2, $3) RETURNING id
+                        VALUES ($1, 2, $2, $3, $4) RETURNING id
                     """,
+                    self._profile_id or 1,
                     item["category"],
                     item["name"],
                     item["value"],
                 )
                 item_id = ins[0][0]
                 if item["tags"]:
-                    await self._conn.executemany(
+                    await self._new_conn.executemany(
                         """
                             INSERT INTO items_tags (item_id, plaintext, name, value)
                             VALUES ($1, $2, $3, $4)
                         """,
                         ((item_id, *tag) for tag in item["tags"]),
                     )
-                await self._conn.execute(
-                    "DELETE FROM items_old WHERE id IN ($1)", del_ids
+                await self._old_conn.execute(
+                    f"DELETE FROM {self._items_table} WHERE id IN ($1)", del_ids
                 )
