+++ acapy_wallet_upgrade/pg_mwst_connection.py
@@ -115,107 +115,5 @@ class PgMWSTConnection(PgConnection):
             """
         )
 
-    def get_wallet(self, old_conn: Connection, wallet_id: str) -> "PgMWSTWallet":
-        return PgMWSTWallet(old_conn, self._conn, wallet_id, "items")
-
-
-class PgMWSTWallet(PgWallet):
-    def __init__(
-        self,
-        old_conn: asyncpg.Connection,
-        new_conn: asyncpg.Connection,
-        wallet_id: str,
-        items_table: str,
-    ):
-        self._old_conn = old_conn
-        self._new_conn = new_conn
-        self._wallet_id = wallet_id
-        self._items_table = items_table
-        self._profile_id = None
-
-    @property
-    def profile_id(self):
-        if not self._profile_id:
-            raise UpgradeError("Profile has not been initialized")
-        return self._profile_id
-
-    async def insert_profile(self, name: str, key: bytes):
-        """Insert the initial profile."""
-        async with self._new_conn.transaction():
-            id_row = await self._new_conn.fetch(
-                """
-                    INSERT INTO profiles (name, profile_key) VALUES($1, $2)
-                    ON CONFLICT DO NOTHING RETURNING id
-                """,
-                name,
-                key,
-            )
-            self._profile_id = id_row[0][0]
-            return self._profile_id
-
-    async def get_metadata(self):
-        stmt = await self._old_conn.fetch(
-            "SELECT value FROM metadata WHERE wallet_id = $1", (self._wallet_id)
-        )
-        found = None
-        if stmt != "":
-            for row in stmt:
-                decoded = base64.b64decode(bytes.decode(row[0]))
-                if found is None:
-                    found = decoded
-                else:
-                    raise Exception("Found duplicate row")
-            return found
-
-        else:
-            raise Exception("Row not found")
-
-    async def fetch_pending_items(self, limit: int):
-        """Fetch un-updated items by wallet_id."""
-        while True:
-            rows = await self._old_conn.fetch(
-                f"""
-                SELECT i.id, i.type, i.name, i.value, i.key,
-                (SELECT string_agg(encode(te.name::bytea, 'hex') || ':' || encode(te.value::bytea, 'hex')::text, ',')
-                    FROM tags_encrypted te WHERE te.item_id = i.id) AS tags_enc,
-                (SELECT string_agg(encode(tp.name::bytea, 'hex') || ':' || encode(tp.value::bytea, 'hex')::text, ',')
-                    FROM tags_plaintext tp WHERE tp.item_id = i.id) AS tags_plain
-                FROM {self._items_table} i WHERE i.wallet_id = $1 LIMIT $2;
-                """,  # noqa
-                self._wallet_id,
-                limit,
-            )
-            if not rows:
-                break
-            yield rows
-
-    async def update_items(self, items):
-        """Update items in the database.
-
-        Differences from PgMWSTWallet:
-        """
-        for item in items:
-            del_ids = item["id"]
-            async with self._new_conn.transaction():
-                ins = await self._new_conn.fetch(
-                    """
-                        INSERT INTO items (profile_id, kind, category, name, value)
-                        VALUES ($1, 2, $2, $3, $4) RETURNING id
-                    """,
-                    self.profile_id,
-                    item["category"],
-                    item["name"],
-                    item["value"],
-                )
-                item_id = ins[0][0]
-                if item["tags"]:
-                    await self._new_conn.executemany(
-                        """
-                            INSERT INTO items_tags (item_id, plaintext, name, value)
-                            VALUES ($1, $2, $3, $4)
-                        """,
-                        ((item_id, *tag) for tag in item["tags"]),
-                    )
-                await self._old_conn.execute(
-                    f"DELETE FROM {self._items_table} WHERE id IN ($1)", del_ids
-                )
+    def get_wallet(self, old_conn: Connection, wallet_id: str) -> "PgWallet":
+        return PgWallet(old_conn, self._conn, "items", wallet_id)
